feat(graphql): modify application.properties, code restructure
